Add draggable icon and make it work
