+++ src/components/shared/List/components/ListItem/index.ts
@@ -1,0 +1,1 @@
+export {default} from "./ListItem"
