+++ src/components/shared/List/components/ListItem/listItem.types.ts
@@ -1,0 +1,10 @@
+export type IListItem = {
+  id: string;
+  [key: string]: any;
+};
+
+export interface IListItemProps {
+  item: IListItem;
+  draggable: boolean;
+  renderValueByKey: keyof IListItem;
+}
